Flower_Classification_Project/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
tensorflow==2.13.0
Pillow==10.0.1
numpy==1.24.3
asyncer==0.0.2
